feat: modify dialog
