+++ at_onboarding_flutter/lib/at_onboarding_otp_screen.dart
@@ -3,6 +3,7 @@ import 'dart:convert';
 import 'package:at_onboarding_flutter/services/size_config.dart';
 import 'package:at_onboarding_flutter/utils/at_onboarding_dimens.dart';
 import 'package:at_onboarding_flutter/utils/color_constants.dart';
+import 'package:at_onboarding_flutter/utils/input_formatter.dart';
 import 'package:at_onboarding_flutter/utils/strings.dart';
 import 'package:at_onboarding_flutter/widgets/at_onboarding_button.dart';
 import 'package:at_onboarding_flutter/widgets/at_onboarding_dialog.dart';
@@ -15,7 +16,6 @@ import 'package:url_launcher/url_launcher.dart';
 import 'at_onboarding_accounts_screen.dart';
 import 'at_onboarding_reference_screen.dart';
 import 'services/free_atsign_service.dart';
-import 'widgets/custom_dialog.dart';
 
 class AtOnboardingOTPResult {
   String atSign;
@@ -134,7 +134,7 @@ class _AtOnboardingOTPScreenState extends State<AtOnboardingOTPScreen> {
                   // controller: _otpController,
                   keyboardType: TextInputType.text,
                   inputFormatters: <TextInputFormatter>[
-                    UpperCaseInputFormatter(),
+                    InputFormatter(),
                   ],
                   onChanged: (String value) {},
                 ),
