+++ at_onboarding_flutter/lib/utils/input_formatter.dart
@@ -1,0 +1,13 @@
+
+import 'package:flutter/services.dart';
+
+class InputFormatter extends TextInputFormatter {
+  @override
+  TextEditingValue formatEditUpdate(
+      TextEditingValue oldValue, TextEditingValue newValue) {
+    return TextEditingValue(
+      text: newValue.text.toUpperCase(),
+      selection: newValue.selection,
+    );
+  }
+}
